backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
passlib[bcrypt]
PyJWT
bcrypt==3.2.2